Split the SyslogTest into multiple tests with different environments.
